Check if model already exists

+++ TrainAndTestGradientQL_LSTM4.py
@@ -7,6 +7,8 @@ import pickle
 import ReinforcementLearning.QlearningFunctions as QL
 import Envs.PytorchEnvironments as Envs
 import torch
+import sys
+import os
 
 q = []
 if torch.cuda.is_available():
@@ -26,12 +28,15 @@ def TrainAndTest(alpha_reward, beta_reward, Tf, Nit, discount_factor, num_episod
     TransEnv = Envs.EnvFeedbackGeneral(Tf, alpha_reward, beta_reward, Channel_Local, batch, M=5)
     TransEnv = TransEnv.to(device)
     model_file = 'ModelCNN_LSTM_Batch_GE_Isolated'+string_alpha+'.pickle'
-    t0 = time.time()
-    Q, policy = Train(TransEnv, discount_factor, num_episodes, epsilon)
-    t1 = time.time()
-    
-    with open('Data/'+model_file, 'wb') as f:
-      pickle.dump(Q, f)    
+    if os.path.isfile('Data/'+model_file):
+      with open('Data/'+model_file, 'rb') as f:
+        Q = pickle.load(f)
+    else:
+      t0 = time.time()
+      Q, policy = Train(TransEnv, discount_factor, num_episodes, epsilon)
+      t1 = time.time()
+      with open('Data/'+model_file, 'wb') as f:
+        pickle.dump(Q, f)
 
     print('Training takes {} seconds'.format(t1 - t0))
     t0 = time.time()
